% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{AllocateIdsRequest}
\alias{AllocateIdsRequest}
\title{AllocateIdsRequest Object}
\usage{
AllocateIdsRequest(keys = NULL)
}
\arguments{
\item{keys}{A list of keys with incomplete key paths for which to allocate IDs}
}
\value{
AllocateIdsRequest object
}
\description{
AllocateIdsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Datastore.AllocateIds.
}
\seealso{
Other AllocateIdsRequest functions: \code{\link{projects.allocateIds}}
}
