% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{AllocateIdsResponse}
\alias{AllocateIdsResponse}
\title{AllocateIdsResponse Object}
\usage{
AllocateIdsResponse(keys = NULL)
}
\arguments{
\item{keys}{The keys specified in the request (in the same order), each with}
}
\value{
AllocateIdsResponse object
}
\description{
AllocateIdsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Datastore.AllocateIds.
}

