% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{ArrayValue}
\alias{ArrayValue}
\title{ArrayValue Object}
\usage{
ArrayValue(values = NULL)
}
\arguments{
\item{values}{Values in the array}
}
\value{
ArrayValue object
}
\description{
ArrayValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An array value.
}

