% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{BeginTransactionRequest}
\alias{BeginTransactionRequest}
\title{BeginTransactionRequest Object}
\usage{
BeginTransactionRequest()
}
\value{
BeginTransactionRequest object
}
\description{
BeginTransactionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Datastore.BeginTransaction.
}
\seealso{
Other BeginTransactionRequest functions: \code{\link{projects.beginTransaction}}
}
