% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{BeginTransactionResponse}
\alias{BeginTransactionResponse}
\title{BeginTransactionResponse Object}
\usage{
BeginTransactionResponse(transaction = NULL)
}
\arguments{
\item{transaction}{The transaction identifier (always present)}
}
\value{
BeginTransactionResponse object
}
\description{
BeginTransactionResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Datastore.BeginTransaction.
}

