% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{CommitRequest}
\alias{CommitRequest}
\title{CommitRequest Object}
\usage{
CommitRequest(transaction = NULL, mode = NULL, mutations = NULL)
}
\arguments{
\item{transaction}{The identifier of the transaction associated with the commit}

\item{mode}{The type of commit to perform}

\item{mutations}{The mutations to perform}
}
\value{
CommitRequest object
}
\description{
CommitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Datastore.Commit.
}
\seealso{
Other CommitRequest functions: \code{\link{projects.commit}}
}
