% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{CommitResponse}
\alias{CommitResponse}
\title{CommitResponse Object}
\usage{
CommitResponse(mutationResults = NULL, indexUpdates = NULL)
}
\arguments{
\item{mutationResults}{The result of performing the mutations}

\item{indexUpdates}{The number of index entries updated during the commit, or zero if none were}
}
\value{
CommitResponse object
}
\description{
CommitResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Datastore.Commit.
}

