% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_functions.R
\docType{package}
\name{datastore_googleAuthR}
\alias{datastore_googleAuthR}
\alias{datastore_googleAuthR-package}
\title{Google Cloud Datastore API
Accesses the schemaless NoSQL database to provide fully managed, robust, scalable storage for your application.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:43:26
filename: /Users/mark/dev/R/autoGoogleAPI/googledatastorev1.auto/R/datastore_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}
}
