% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{EntityResult}
\alias{EntityResult}
\title{EntityResult Object}
\usage{
EntityResult(cursor = NULL, entity = NULL, version = NULL)
}
\arguments{
\item{cursor}{A cursor that points to the position after the result entity}

\item{entity}{The resulting entity}

\item{version}{The version of the entity, a strictly positive number that monotonically}
}
\value{
EntityResult object
}
\description{
EntityResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of fetching an entity from Datastore.
}
