% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Filter}
\alias{Filter}
\title{Filter Object}
\usage{
Filter(propertyFilter = NULL, compositeFilter = NULL)
}
\arguments{
\item{propertyFilter}{A filter on a property}

\item{compositeFilter}{A composite filter}
}
\value{
Filter object
}
\description{
Filter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A holder for any type of filter.
}
