% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{GqlQuery.namedBindings}
\alias{GqlQuery.namedBindings}
\title{GqlQuery.namedBindings Object}
\usage{
GqlQuery.namedBindings()
}
\value{
GqlQuery.namedBindings object
}
\description{
GqlQuery.namedBindings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
For each non-reserved named binding site in the query string, there must bea named parameter with that name, but not necessarily the inverse.Key must match regex `A-Za-z_$*`, must not match regex`__.*__`, and must not be `''`.
}
\seealso{
Other GqlQuery functions: \code{\link{GqlQuery}}
}
