% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{GqlQuery}
\alias{GqlQuery}
\title{GqlQuery Object}
\usage{
GqlQuery(GqlQuery.namedBindings = NULL, queryString = NULL,
  namedBindings = NULL, allowLiterals = NULL, positionalBindings = NULL)
}
\arguments{
\item{GqlQuery.namedBindings}{The \link{GqlQuery.namedBindings} object or list of objects}

\item{queryString}{A string of the format described}

\item{namedBindings}{For each non-reserved named binding site in the query string, there must be}

\item{allowLiterals}{When false, the query string must not contain any literals and instead must}

\item{positionalBindings}{Numbered binding site @1 references the first numbered parameter,}
}
\value{
GqlQuery object
}
\description{
GqlQuery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A [GQL query](https://cloud.google.com/datastore/docs/apis/gql/gql_reference).
}
\seealso{
Other GqlQuery functions: \code{\link{GqlQuery.namedBindings}}
}
