% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{GqlQueryParameter}
\alias{GqlQueryParameter}
\title{GqlQueryParameter Object}
\usage{
GqlQueryParameter(value = NULL, cursor = NULL)
}
\arguments{
\item{value}{A value parameter}

\item{cursor}{A query cursor}
}
\value{
GqlQueryParameter object
}
\description{
GqlQueryParameter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A binding parameter for a GQL query.
}
