% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{LookupRequest}
\alias{LookupRequest}
\title{LookupRequest Object}
\usage{
LookupRequest(readOptions = NULL, keys = NULL)
}
\arguments{
\item{readOptions}{The options for this lookup request}

\item{keys}{Keys of entities to look up}
}
\value{
LookupRequest object
}
\description{
LookupRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Datastore.Lookup.
}
\seealso{
Other LookupRequest functions: \code{\link{projects.lookup}}
}
