% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{MutationResult}
\alias{MutationResult}
\title{MutationResult Object}
\usage{
MutationResult(conflictDetected = NULL, key = NULL, version = NULL)
}
\arguments{
\item{conflictDetected}{Whether a conflict was detected for this mutation}

\item{key}{The automatically allocated key}

\item{version}{The version of the entity on the server after processing the mutation}
}
\value{
MutationResult object
}
\description{
MutationResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of applying a mutation.
}

