% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PathElement}
\alias{PathElement}
\title{PathElement Object}
\usage{
PathElement(id = NULL, name = NULL)
}
\arguments{
\item{id}{The auto-allocated ID of the entity}

\item{name}{The name of the entity}
}
\value{
PathElement object
}
\description{
PathElement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A (kind, ID/name) pair used to construct a key path.If either name or ID is set, the element is complete.If neither is set, the element is incomplete.
}
