% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Query}
\alias{Query}
\title{Query Object}
\usage{
Query(limit = NULL, filter = NULL, endCursor = NULL, distinctOn = NULL,
  offset = NULL, projection = NULL, order = NULL, startCursor = NULL)
}
\arguments{
\item{limit}{The maximum number of results to return}

\item{filter}{The filter to apply}

\item{endCursor}{An ending point for the query results}

\item{distinctOn}{The properties to make distinct}

\item{offset}{The number of results to skip}

\item{projection}{The projection to return}

\item{order}{The order to apply to the query results (if empty, order is unspecified)}

\item{startCursor}{A starting point for the query results}
}
\value{
Query object
}
\description{
Query Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A query for entities.
}
