% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{QueryResultBatch}
\alias{QueryResultBatch}
\title{QueryResultBatch Object}
\usage{
QueryResultBatch(snapshotVersion = NULL, endCursor = NULL,
  skippedCursor = NULL, entityResultType = NULL, moreResults = NULL,
  entityResults = NULL, skippedResults = NULL)
}
\arguments{
\item{snapshotVersion}{The version number of the snapshot this batch was returned from}

\item{endCursor}{A cursor that points to the position after the last result in the batch}

\item{skippedCursor}{A cursor that points to the position after the last skipped result}

\item{entityResultType}{The result type for every entity in `entity_results`}

\item{moreResults}{The state of the query after the current batch}

\item{entityResults}{The results for this batch}

\item{skippedResults}{The number of results skipped, typically because of an offset}
}
\value{
QueryResultBatch object
}
\description{
QueryResultBatch Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch of results produced by a query.
}
