% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{ReadOptions}
\alias{ReadOptions}
\title{ReadOptions Object}
\usage{
ReadOptions(transaction = NULL, readConsistency = NULL)
}
\arguments{
\item{transaction}{The identifier of the transaction in which to read}

\item{readConsistency}{The non-transactional read consistency to use}
}
\value{
ReadOptions object
}
\description{
ReadOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The options shared by read requests.
}
