% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{RunQueryRequest}
\alias{RunQueryRequest}
\title{RunQueryRequest Object}
\usage{
RunQueryRequest(partitionId = NULL, gqlQuery = NULL, readOptions = NULL,
  query = NULL)
}
\arguments{
\item{partitionId}{Entities are partitioned into subsets, identified by a partition ID}

\item{gqlQuery}{The GQL query to run}

\item{readOptions}{The options for this query}

\item{query}{The query to run}
}
\value{
RunQueryRequest object
}
\description{
RunQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Datastore.RunQuery.
}
\seealso{
Other RunQueryRequest functions: \code{\link{projects.runQuery}}
}

