% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{RunQueryResponse}
\alias{RunQueryResponse}
\title{RunQueryResponse Object}
\usage{
RunQueryResponse(batch = NULL, query = NULL)
}
\arguments{
\item{batch}{A batch of query results (always present)}

\item{query}{The parsed form of the `GqlQuery` from the request, if it was set}
}
\value{
RunQueryResponse object
}
\description{
RunQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Datastore.RunQuery.
}
