% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Value}
\alias{Value}
\title{Google Cloud Datastore API Objects 
Accesses the schemaless NoSQL database to provide fully managed, robust, scalable storage for your application.}
\usage{
Value(stringValue = NULL, arrayValue = NULL, entityValue = NULL,
  meaning = NULL, integerValue = NULL, doubleValue = NULL,
  blobValue = NULL, geoPointValue = NULL, nullValue = NULL,
  booleanValue = NULL, keyValue = NULL, excludeFromIndexes = NULL,
  timestampValue = NULL)
}
\arguments{
\item{stringValue}{A UTF-8 encoded string value}

\item{arrayValue}{An array value}

\item{entityValue}{An entity value}

\item{meaning}{The `meaning` field should only be populated for backwards compatibility}

\item{integerValue}{An integer value}

\item{doubleValue}{A double value}

\item{blobValue}{A blob value}

\item{geoPointValue}{A geo point value representing a point on the surface of Earth}

\item{nullValue}{A null value}

\item{booleanValue}{A boolean value}

\item{keyValue}{A key value}

\item{excludeFromIndexes}{If the value should be excluded from all indexes including those defined}

\item{timestampValue}{A timestamp value}
}
\value{
Value object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:05:05
filename: /Users/mark/dev/R/autoGoogleAPI/googledatastorev1.auto/R/datastore_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Value Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message that can hold any of the supported value types and associatedmetadata.
}

